% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse a vector sequence}
\usage{
collapse(x)
}
\arguments{
\item{x}{a vector sequence}
}
\value{
A character sequence
}
\description{
Collapse a vector sequence
}
\section{Further details}{

Converts a vector sequence to a character sequence
}
\examples{
collapse(LETTERS[1:10])
}

