% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_seq.R
\name{get_seq}
\alias{get_seq}
\title{Get behavioral sequence from a dataframe of behavioral observations}
\usage{
get_seq(df, begin = 0, end = 1200, varx = "state", vary = "time",
  drop = F, collapse = T)
}
\arguments{
\item{df}{dataframe containing behavioral observations}

\item{begin}{time at which to start sequence}

\item{end}{time at which to end sequence}

\item{varx}{character defining name of column to make into a sequence}

\item{vary}{character defining name of column containing time values}

\item{drop}{whether to include repeated behavior codes}

\item{collapse}{whether to make sequence a character sequence (default)
or a  vector sequence (if collpase=F)}
}
\value{
A character of vector sequence
}
\description{
Get behavioral sequence from a dataframe of behavioral observations
}
\section{Further details}{

Get behavioral sequence from a dataframe of behavioral observations
}
\examples{
get_seq(mouse, varx = 'stateX', begin=1100, end=1200)
get_seq(mouse, varx = 'stateX', drop=T, begin=1100, end=1200)
}

