% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_seq0.R
\name{get_seq0}
\alias{get_seq0}
\title{Get timed-interval behavioral sequence from a dataframe of behavioral observations}
\usage{
get_seq0(df, begin = 1, end = 1199, interval = 1, varx = "state",
  vary = "time", collapse = T)
}
\arguments{
\item{df}{dataframe containing behavioral observations}

\item{begin}{time at which to start sequence}

\item{end}{time at which to end sequence}

\item{interval}{the time unit inteval over which to add successive codes}

\item{varx}{character defining name of column to make into a sequence}

\item{vary}{character defining name of column containing time values}

\item{collapse}{whether to make sequence a character sequence (default)
or a  vector sequence (if collpase=F)}
}
\value{
A character of vector sequence
}
\description{
Get timed-interval behavioral sequence from a dataframe of behavioral observations
}
\section{Further details}{

Get timed-interval  behavioral sequence from a dataframe of behavioral observations
}
\examples{
get_seq0(mouse, varx = 'behaviorX', begin=1150, end=1200, interval=1)
}

