% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mat_trans.R
\name{mat_trans}
\alias{mat_trans}
\title{Get transition probabilities from observed frequency matrix}
\usage{
mat_trans(m)
}
\arguments{
\item{m}{a matrix or table of observed transition frequencies}
}
\value{
Table of observed frequencies
}
\description{
Get transition probabilities from observed frequency matrix
}
\section{Further details}{

Get transition probabilities from observed frequency matrix
}
\examples{
a<-seq_obs("ababababcdcdbabdcbadbcadbcadcb")
mat_trans(a)
}

