% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob_con.R
\name{prob_con}
\alias{prob_con}
\title{Get simple conditional probailities of behaviors}
\usage{
prob_con(x)
}
\arguments{
\item{x}{a character sequence}
}
\value{
A vector of conditional probabilities
}
\description{
Get simple conditional probailities of behaviors
}
\section{Further details}{

Calculates simple conditional probabilties of behaviors in
a character sequence at a lag of 1.
}
\examples{
prob_con("abacbabacbabcababacbcbcbabaaabababbababcb")
prob_con("AABABABBABAB")
prob_con(jaffe) #on of the example sequences
}

