% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob_ex.R
\name{prob_ex}
\alias{prob_ex}
\title{Get unconditional probability of a behavior}
\usage{
prob_ex(x)
}
\arguments{
\item{x}{a character sequence}
}
\value{
A vector of unconditional probabilities
}
\description{
Get unconditional probability of a behavior
}
\section{Further details}{

Calculates simple unconditional probabilities of behaviors in
a character sequence
}
\examples{
prob_ex("abacbabacbabcababacbcbcbabaaabababbababcb")
}

