% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seq_clip.R
\name{seq_clip}
\alias{seq_clip}
\title{Extract a sequence at specific points}
\usage{
seq_clip(x, b = 1, e = 1)
}
\arguments{
\item{x}{a character sequence}

\item{b}{the beginning position of extracted sequence}

\item{e}{the end position of extracted sequence}
}
\value{
An extracted subsequence of original sequence
}
\description{
Extract a sequence at specific points
}
\section{Further details}{

Clip a sequence to extract a subsequence
}
\examples{
seq_clip("abacbabacbabcababacbcbcbabaaabababbababcb", b=4,e=10)
}

