% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seq_droprpt.R
\name{seq_droprpt}
\alias{seq_droprpt}
\title{Drop repeated elements from a sequence}
\usage{
seq_droprpt(x, coll = F)
}
\arguments{
\item{x}{a vector or character sequence}

\item{coll}{whether input sequence is a
vector sequence  \strong{\code{coll}=F} or a
character sequence  \strong{\code{coll}=T}}
}
\value{
A character or vector sequence with no
repeated elements
}
\description{
Drop repeated elements from a sequence
}
\section{Further details}{

Drops repeating elements from a vector or
character sequence
}
\examples{
seq_droprpt(c("A", "A", "A", "B", "C","C", "C"))
seq_droprpt('AAABBBCCCDDDEEEBBBAAA', coll=T)
}

