% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seq_exp.R
\name{seq_exp}
\alias{seq_exp}
\title{Get expected frequencies of first order transitions}
\usage{
seq_exp(x)
}
\arguments{
\item{x}{a character sequence}
}
\value{
Table of observed frequencies
}
\description{
Get expected frequencies of first order transitions
}
\section{Further details}{

Generates the expected frequencies of first order
transitions.  Note: this shouldn't be used in there
are structural zeros in the matrix of observed
transition frequencies.
}
\examples{
seq_exp("abacbabacbabcababacbcbcbabaaabababbababcb")
}

