% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seq_exp0.R
\name{seq_exp0}
\alias{seq_exp0}
\title{Get expected frequencies of first order transitions for
non-repeating sequences}
\usage{
seq_exp0(x)
}
\arguments{
\item{x}{a character sequence with repeated behaviors not allowed}
}
\value{
Table of observed frequencies
}
\description{
Get expected frequencies of first order transitions for
non-repeating sequences
}
\section{Further details}{

This function uses iterative proportional fitting to find expected frequencies
for non-repeating  sequences
}
\examples{
seq_exp0(seq_droprpt("abacbabacbabcababacbcbcbabaaabababbababcb",coll=T))
}

