% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seq_obs.R
\name{seq_obs}
\alias{seq_obs}
\title{Get observed frequencies of transitions}
\usage{
seq_obs(x, l = 1)
}
\arguments{
\item{x}{a character sequence}

\item{l}{the lag}
}
\value{
Table of observed frequencies
}
\description{
Get observed frequencies of transitions
}
\section{Further details}{

Generates a table of frequencies of observed transitions
at specified lag
}
\examples{
seq_obs("abacbabacbabcababacbcbcbabaaabababbababcb")
seq_obs("abacbabacbabcababacbcbcbabaaabababbababcb", 2)
}

