% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seq_rndtest.R
\name{seq_rndtest}
\alias{seq_rndtest}
\title{Randomization test of behavioral contingencies in a sequence}
\usage{
seq_rndtest(seqz, nx = 2, nperms = 10000)
}
\arguments{
\item{nx}{length of substrings}

\item{nperms}{number of permutations}

\item{x}{a character sequence}
}
\value{
A vector of probabilities for each contingency
}
\description{
Randomization test of behavioral contingencies in a sequence
}
\section{Further details}{

Performs a randomization test to determine which contingencies
occur at an observed frequency higher than chance.
}
\examples{
seq_rndtest("abacbabacbabcababacbcbcbabaaabababbababcb",2)
seq_rndtest("abacbabacbabcababacbcbcbabaaabababbababcb",3)
}

