% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seq_split.R
\name{seq_split}
\alias{seq_split}
\title{Split a sequence into roughly equal parts}
\usage{
seq_split(x, n = 2)
}
\arguments{
\item{x}{a character sequence}

\item{n}{the number of parts to split the sequence into}
}
\value{
A list of split character sequences
}
\description{
Split a sequence into roughly equal parts
}
\section{Further details}{

Split a character sequence into roughly equal parts
}
\examples{
seq_split("abacbabacbabcababacbcbcbabaaabababbababcb", 2)
seq_split("abacbabacbabcababacbcbcbabaaabababbababcb", 3)
}

