% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seq_substrings.R
\name{seq_substrings}
\alias{seq_substrings}
\title{Get all subsequences of a given length}
\usage{
seq_substrings(x, n = 2)
}
\arguments{
\item{x}{a character sequence}

\item{n}{length of substrings}
}
\value{
A vector of all possible substrings
}
\description{
Get all subsequences of a given length
}
\section{Further details}{

Calculates all possible substrings of a given length in
a character sequence
}
\examples{
seq_substrings("abacbabacbabcababacbcbcbabaaabababbababcb")
}

