% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqfreq.R
\name{seqfreq}
\alias{seqfreq}
\title{Get observed frequncy of a behavior}
\usage{
seqfreq(x)
}
\arguments{
\item{x}{a character sequence}
}
\value{
A vector of observed frequencies
}
\description{
Get observed frequncy of a behavior
}
\section{Further details}{

Calculates simple observed frequencies of behaviors in
a character sequence
}
\examples{
seqfreq("abacbabacbabcababacbcbcbabaaabababbababcb")
}

