% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subseq_lcs.R
\name{subseq_lcs}
\alias{subseq_lcs}
\title{Find the longest common subsequence to two sequences}
\usage{
subseq_lcs(x, y, type = "df")
}
\arguments{
\item{x}{character sequence 1}

\item{y}{character sequence 2}

\item{\strong{\code{type}="df"}}{return a dataframe
\strong{\code{type}="v"} return a vector}
}
\value{
Longest common subsequence in a dataframe with
indices of each behavior in x and y in columns or in a vector
}
\description{
Find the longest common subsequence to two sequences
}
\section{Further details}{

Find the longest common subsequence to two sequences.  The two
types utilize different methods and may produce varying solutions -
there is not necessarily one unique solution
}
\examples{
subseq_lcs('asdfasdfasdfasdf', 'asdgdsfagasdfsg')
}

