% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subseq_long.R
\name{subseq_long}
\alias{subseq_long}
\title{Find length of longest subsequence common to two sequences}
\usage{
subseq_long(x, y)
}
\arguments{
\item{x}{character sequence 1}

\item{y}{character sequence 2}
}
\value{
Length of longest subsequence common to both
}
\description{
Find length of longest subsequence common to two sequences
}
\section{Further details}{

Find length of longest subsequence common to two sequences
}
\examples{
subseq_long("cacaabdcabdcab", "andcabdcadcad")
}

