% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subseq_long0.R
\name{subseq_long0}
\alias{subseq_long0}
\title{Find the longest subsequence(s) common to two sequences}
\usage{
subseq_long0(x, y)
}
\arguments{
\item{x}{character sequence 1}

\item{y}{character sequence 2}
}
\value{
Longest subsequence(s) common to both
}
\description{
Find the longest subsequence(s) common to two sequences
}
\section{Further details}{

Find the longest subsequence(s) common to two sequences
}
\examples{
subseq_long0("ababdcbdabdbcabadbcbabbcbdabc", "dbadbcdbadbcaddbdddbc")
}

