% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subseq_tot.R
\name{subseq_tot}
\alias{subseq_tot}
\title{Find indices of subsequences in a given subsequence}
\usage{
subseq_tot(x, y)
}
\arguments{
\item{x}{subsequence to test for}

\item{y}{sequence to compare to}
}
\value{
Total number of instances of subsequence,
the indices of the subsequences in the sequence,
the length of the sequence, the frequency of the
subsequence
}
\description{
Find indices of subsequences in a given subsequence
}
\section{Further details}{

Find indices of subsequences in a given subsequence
}
\examples{
subseq_tot("ab","adcabdadabdadabdabd")
}

