% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uncollapse.R
\name{uncollapse}
\alias{uncollapse}
\title{Uncollapse a vector sequence}
\usage{
uncollapse(x, n = 1)
}
\arguments{
\item{x}{a character sequence}

\item{n}{The number of codes to put into each distinct
element.}
}
\value{
A vector sequence
}
\description{
Uncollapse a vector sequence
}
\section{Further details}{

Converts a character sequence to a vector sequence
}
\examples{
uncollapse("ABCDEFGABCDEFG")
}

